Add CI to Website!
- Automated testing of the entires site (No more broken links)
- Fixes the {{site.base}} problems
- On release (Commit to master), builds, tests and updates a release
containing the files need to deploy on AWPS
- Adds make deploy-github which fetches the latest release and copies it to AWPS

This does mean protected branches should be enabled for master. As master
is now treated as the "released" version of the site.
- Enable protected branches for master
- Require status checks to pass before merging: Enable
- Require pull request reviews before merging: Enable, Incepts requires 1 review
- Include Administrators: Enable

Quick Question: I assumed the preview staging server on AWPS is /collections/aced is that correct?

1) User makes an edit, commits and pushes
2) GitHub Actions runs unit tests, and reports status
- If everything passes -> Ready for review
- Otherwise more edits are needed (Log files will show what went wrong)
3) Someone approves and merges the pull request
4) GitHub Actions runs unit tests again, and updates released files (if test pass)
5) Writer on AWPS runs 'make deploy-github` to transfer files to AWPS, and then publishes on AWPS

Note: This also switches from rsync (Doesn't support SFTP) to rclone (Does), giving us
command-line access to AWPS. Make assumes cmu-awps is the rclone host